separating build libs - improved cmake file - new module example
