+++ example/example-video-ts.html
@@ -24,7 +24,7 @@
         var canvasWidth;
         var canvasHeight;
         var vec, conf, corners, markerPos;
-        var ar;
+        var trk;
         var videoSize;
 
         if (navigator.mediaDevices && navigator.mediaDevices.getUserMedia) {
@@ -64,15 +64,17 @@
                     canvas.width = canvasWidth;
                     canvas.height = canvasHeight;
                     if (video) {
-                        ARToolKitPlus.initTrackerSingleMarker(false, "data/PGR_M12x0.5_2.5mm.cal", width, height, 80).then(
-                            tracker => {
-                                console.log(tracker);
-                                ar = tracker;
-                                // here we go, just two calls to find the camera pose
-                                console.log("PIXEL_FORMAT: ", ARToolKitPlus.PIXEL_FORMAT.PIXEL_FORMAT_LUM)
-                                createOverlayCanvas()
-                                process();
-                            })
+                        ARToolKitPlus.ARToolkitPlus.initAR().then((ar) => {
+                            ARToolKitPlus.TrackerSingleMarker.initTrackerSingleMarker(false, "data/PGR_M12x0.5_2.5mm.cal", width, height, 80).then(
+                                tracker => {
+                                    console.log(tracker);
+                                    trk = tracker;
+                                    // Display the Pixel Format in use, just to test the static variables...
+                                    console.log("PIXEL_FORMAT: ", ARToolKitPlus.ARToolkitPlus.PIXEL_FORMAT.PIXEL_FORMAT_LUM);
+                                    createOverlayCanvas()
+                                    process();
+                                })
+                        })
                     }
 
                 });
@@ -145,25 +147,25 @@
                 overlayCtx.font = "10px serif";
                 overlayCtx.fillStyle = 'red';
                 overlayCtx.fill();
-                overlayCtx.fillText(ar.getMarkerId(), markerPos.x + 10, markerPos.y + 10);
+                overlayCtx.fillText(trk.getMarkerId(), markerPos.x + 10, markerPos.y + 10);
                 overlayCtx.fillStyle = 'green';
                 overlayCtx.fill();
             }
 
             function process() {
-                corners = ar.getMarkerVertexes();
+                corners = trk.getMarkerVertexes();
                 // Print in the console the vertexes (corners) of the marker rect
                 console.log("corners: ", corners);
-                markerPos = ar.getMarkerPos();
+                markerPos = trk.getMarkerPos();
                 ctx.drawImage(video, 0, 0, width, height);
                 var image_data = ctx.getImageData(0, 0, width, height);
 
                 if (image_data.data) {
-                    vec = ar.update(grayscale(image_data.data));
+                    vec = trk.update(grayscale(image_data.data));
                     console.log("Marker found is: ", vec.get(0));
-                    conf = ar.getConfidence();
+                    conf = trk.getConfidence();
                     console.log("Marker confidence is: ", conf * 100, "%");
-                    var matrix = ar.getModelViewMatrix();
+                    var matrix = trk.getModelViewMatrix();
                     console.log("Marker pose matrix is:", matrix);
                     ctx.putImageData(image_data, 0, 0);
                 }
