+++ example/example.html
@@ -10,6 +10,7 @@
 
 <body>
     <script src="../build/artoolkitplus_em_debug.js"></script>
+    <script src="../build/trackerSM_debug.js"></script>
     <script src="loader.js"></script>
     <script>
         console.log(Module);
