+++ js/trackerSM_post.api.js
@@ -1,0 +1,4 @@
+Module["onRuntimeInitialized"] = function() {
+    var event = new Event('trackerSM-loaded');
+    window.dispatchEvent(event);
+}
